Fix to work with custom fs

+++ rmrf.js
@@ -1,13 +1,27 @@
 const rimraf = require('rimraf');
 
 module.exports = function (fs, p) {
-  return new Promise((resolve, reject) => {
-    rimraf(p, fs, err => {
-      if (err) {
-        return reject(err);
+  return new Promise(async (resolve, reject) => {
+    try {
+      if (fs.rmrf) {
+        await fs.rmrf(p, err => {
+          if (err) {
+            return reject(err);
+          }
+          resolve();
+        });
+        return;
       }
 
-      resolve();
-    });
+      rimraf(p, fs, err => {
+        if (err) {
+          return reject(err);
+        }
+
+        resolve();
+      });
+    } catch (err) {
+      reject(err);
+    }
   });
 };
